Rename Api.js to api.js
As it is not a class
